Fix int overflow issues for zoom level > 15

+++ src/main/java/ch/poole/geo/pmtiles/Hilbert.java
@@ -3,16 +3,15 @@ package ch.poole.geo.pmtiles;
 /**
  * See https://en.wikipedia.org/wiki/Hilbert_curve#Applications_and_mapping_algorithms
  * 
- * In practical terms this doesn't seem to be slower than 
- * <a href="https://github.com/rawrunprotected/hilbert_curves">github.com/rawrunprotected/hilbert_curves</a> 
- * <a href="https://threadlocalmutex.com/">threadlocalmutex.com</a>
- * but works for all zoom levels.
+ * In practical terms this doesn't seem to be slower than
+ * <a href="https://github.com/rawrunprotected/hilbert_curves">github.com/rawrunprotected/hilbert_curves</a>
+ * <a href="https://threadlocalmutex.com/">threadlocalmutex.com</a> but works for all zoom levels.
  * 
  * https://github.com/davidmoten/hilbert-curve might be an alternative
  *
  */
 public final class Hilbert {
-    
+
     private Hilbert() {
         throw new IllegalStateException("Utility class, can't be instantiated");
     }
@@ -25,11 +24,11 @@ public final class Hilbert {
      * @param y tile y coordinate
      * @return an index along the Hilbert curve
      */
-    public static int zxyToIndex(int z, int x, int y) {
-        int n = 1 << z;
+    public static long zxyToIndex(int z, long x, long y) {
+        long n = 1L << z;
         int rx;
         int ry;
-        int s;
+        long s;
         int d = 0;
         for (s = n / 2; s > 0; s /= 2) {
             rx = (x & s) > 0 ? 1 : 0;
@@ -40,7 +39,7 @@ public final class Hilbert {
                     x = n - 1 - x;
                     y = n - 1 - y;
                 }
-                int t = x;
+                long t = x;
                 x = y;
                 y = t;
             }
